+++ GeneralsRust/Code/GameEngine/GameClient/src/gui/callbacks/popup_player_info.rs
@@ -21,8 +21,8 @@ use game_network::gamespy::buddy_thread::{
 };
 use game_network::gamespy::peer_defs::get_gamespy_info;
 
-const KEY_ESC: u32 = 0x1B;
-const KEY_STATE_UP: u32 = 0x0001;
+const KEY_ESC: usize = 0x1B;
+const KEY_STATE_UP: usize = 0x0001;
 
 #[derive(Default)]
 struct PopupPlayerInfoState {
